cleanup

+++ cbase/fmt.c
@@ -4452,7 +4452,7 @@ fmt_float32_shortest(char *buffer, int64 capacity, float value) {
         return status;
     }
 
-    len = (int32)f2s_buffered_n(value, temp);
+    len = f2s_buffered_n(value, temp);
     return fmt_float_copy(buffer, capacity, temp, len);
 }
 
@@ -4466,7 +4466,7 @@ fmt_float64_shortest(char *buffer, int64 capacity, double value) {
         return status;
     }
 
-    len = (int32)d2s_buffered_n(value, temp);
+    len = d2s_buffered_n(value, temp);
     return fmt_float_copy(buffer, capacity, temp, len);
 }
 
@@ -4483,7 +4483,7 @@ fmt_float64_fixed(char *buffer, int64 capacity, double value, int32 precision) {
         return status;
     }
 
-    len = (int32)d2fixed_buffered_n(value, (uint32)precision, temp);
+    len = d2fixed_buffered_n(value, (uint32)precision, temp);
     return fmt_float_copy(buffer, capacity, temp, len);
 }
 
@@ -4501,7 +4501,7 @@ fmt_float64_scientific(char *buffer, int64 capacity,
         return status;
     }
 
-    len = (int32)d2exp_buffered_n(value, (uint32)precision, temp);
+    len = d2exp_buffered_n(value, (uint32)precision, temp);
     return fmt_float_copy(buffer, capacity, temp, len);
 }
 
@@ -4510,7 +4510,7 @@ sb_float64(StrBuilder *str_builder, double value) {
     int32 len;
 
     sb_reserve(str_builder, FMT_FLOAT_RYU_BUFFER_SIZE);
-    len = (int32)d2s_buffered_n(value, str_builder->data + str_builder->len);
+    len = d2s_buffered_n(value, str_builder->data + str_builder->len);
     str_builder->len += len;
     str_builder->data[str_builder->len] = '\0';
     return;
